+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/extensions/internal/PollExtensionsTests.kt
@@ -100,6 +100,7 @@ internal class PollExtensionsTests {
             cid = "channel1",
             channelType = "messaging",
             channelId = "channel1",
+            messageId = "message1",
             poll = basePoll,
             newVote = newVote,
         )
@@ -123,6 +124,7 @@ internal class PollExtensionsTests {
             cid = "channel1",
             channelType = "messaging",
             channelId = "channel1",
+            messageId = "message1",
             poll = basePoll,
             newVote = newVote,
         )
@@ -153,6 +155,7 @@ internal class PollExtensionsTests {
             cid = "channel1",
             channelType = "messaging",
             channelId = "channel1",
+            messageId = "message1",
             poll = basePoll,
             newVote = newVote,
         )
@@ -176,6 +179,7 @@ internal class PollExtensionsTests {
             cid = "channel1",
             channelType = "messaging",
             channelId = "channel1",
+            messageId = "message1",
             poll = basePoll,
             removedVote = vote1,
         )
@@ -206,6 +210,7 @@ internal class PollExtensionsTests {
             cid = "channel1",
             channelType = "messaging",
             channelId = "channel1",
+            messageId = "message1",
             poll = basePoll,
             newAnswer = newAnswer,
         )
@@ -229,6 +234,7 @@ internal class PollExtensionsTests {
             cid = "channel1",
             channelType = "messaging",
             channelId = "channel1",
+            messageId = "message1",
             poll = basePoll,
         )
 
@@ -249,6 +255,7 @@ internal class PollExtensionsTests {
             cid = "channel1",
             channelType = "messaging",
             channelId = "channel1",
+            messageId = "message1",
             poll = basePoll,
         )
 
@@ -270,6 +277,7 @@ internal class PollExtensionsTests {
             cid = "channel1",
             channelType = "messaging",
             channelId = "channel1",
+            messageId = "message1",
             poll = basePoll,
             newVote = vote1,
         )