+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/state/plugin/logic/channel/thread/internal/ThreadLogic.kt
@@ -16,15 +16,25 @@
 
 package io.getstream.chat.android.state.plugin.logic.channel.thread.internal
 
+import io.getstream.chat.android.client.events.AnswerCastedEvent
 import io.getstream.chat.android.client.events.HasMessage
+import io.getstream.chat.android.client.events.HasPoll
 import io.getstream.chat.android.client.events.HasReminder
 import io.getstream.chat.android.client.events.MessageUpdatedEvent
+import io.getstream.chat.android.client.events.PollClosedEvent
+import io.getstream.chat.android.client.events.PollDeletedEvent
+import io.getstream.chat.android.client.events.PollUpdatedEvent
 import io.getstream.chat.android.client.events.ReminderCreatedEvent
 import io.getstream.chat.android.client.events.ReminderDeletedEvent
 import io.getstream.chat.android.client.events.ReminderUpdatedEvent
+import io.getstream.chat.android.client.events.VoteCastedEvent
+import io.getstream.chat.android.client.events.VoteChangedEvent
+import io.getstream.chat.android.client.events.VoteRemovedEvent
+import io.getstream.chat.android.client.extensions.internal.processPoll
 import io.getstream.chat.android.client.extensions.internal.toMessageReminderInfo
 import io.getstream.chat.android.client.plugin.listeners.ThreadQueryListener
 import io.getstream.chat.android.models.Message
+import io.getstream.chat.android.models.Poll
 import io.getstream.chat.android.state.plugin.state.channel.thread.internal.ThreadMutableState
 
 /** Logic class for thread state management. Implements [ThreadQueryListener] as listener for LLC requests. */
@@ -130,6 +140,31 @@ internal class ThreadLogic(
         }
         if (messages.isNotEmpty()) {
             upsertMessages(messages)
+        }
+    }
+
+    internal fun handlePollEvents(currentUserId: String?, events: List<HasPoll>) {
+        // Don't handle poll events if there is no poll in the parent message (should never happen)
+        val parentMessage = mutableState.parentMessage ?: return
+        val poll = parentMessage.poll ?: return
+        // The processed poll after applying each event sequentially
+        var processedPoll: Poll? = poll
+        // Don't handle poll events if the poll in the parent message is different (should never happen)
+        events
+            .filter { it.poll.id == poll.id }
+            .forEach { event ->
+                processedPoll = when (event) {
+                    is AnswerCastedEvent -> event.processPoll { processedPoll }
+                    is PollClosedEvent -> event.processPoll { processedPoll }
+                    is PollUpdatedEvent -> event.processPoll { processedPoll }
+                    is VoteRemovedEvent -> event.processPoll { processedPoll }
+                    is VoteCastedEvent -> event.processPoll(currentUserId) { processedPoll }
+                    is VoteChangedEvent -> event.processPoll(currentUserId) { processedPoll }
+                    is PollDeletedEvent -> null // poll is deleted, remove from state
+                }
+            }
+        if (processedPoll != poll) {
+            mutableState.updateParentMessagePoll(processedPoll)
         }
     }
 }